% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jabar_basemap}
\alias{jabar_basemap}
\title{Basemap of Jawa Barat}
\format{An sf object with 27 features and 4 fields.
\describe{
  \item{name_bps}{Name of the area according to BPS}
  \item{code_bps}{Code of the area according to BPS}
  \item{name_kemendagri}{Name of the area according to KEMENDAGRI}
  \item{code_kemendagri}{Code of the area according to KEMENDAGRI}
}}
\usage{
jabar_basemap
}
\description{
Map of district area (Kota/Kabupaten) in Jawa Barat.
}
\keyword{datasets}
