% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabr-package.R
\docType{package}
\name{jabr-package}
\alias{jabr}
\alias{jabr-package}
\title{jabr: Easy Access for West Java Data}
\description{
An R interface to browse, access, and fetch data
    about West Java.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aswansyahputra/jabr}
  \item Report bugs at \url{https://github.com/aswansyahputra/jabr/issues}
}

}
\author{
\strong{Maintainer}: Muhammad Aswan Syahputra \email{muhammadaswansyahputra@gmail.com}

}
\keyword{internal}
