% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabr_basemap.R
\name{jabr_basemap}
\alias{jabr_basemap}
\title{Get basemap}
\usage{
jabr_basemap(level)
}
\arguments{
\item{level}{Level of administrative division. Valid values are "district" for Kota/Kabupaten, "subdistrict" for Kecamatan and "village" for Desa/Kelurahan.}
}
\value{
A tibble.
}
\description{
List basemap of the administrative area in Jawa Barat.
}
\examples{
\donttest{
library(jabr)

jabr_basemap(level = "district")
}

}
