% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jabr_district}
\alias{jabr_district}
\title{Administrative district of Jawa Barat}
\format{A tibble with 27 rows and 4 columns.
\describe{
  \item{name_bps}{Name of the area according to BPS}
  \item{code_bps}{Code of the area according to BPS}
  \item{name_kemendagri}{Name of the area according to KEMENDAGRI}
  \item{code_kemendagri}{Code of the area according to KEMENDAGRI}
}}
\usage{
jabr_district
}
\description{
A list of district area (Kota/Kabupaten) in Jawa Barat.
}
\keyword{datasets}
