% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabr_list_dataset.R
\name{jabr_list_dataset}
\alias{jabr_list_dataset}
\title{List Dataset}
\usage{
jabr_list_dataset(update = FALSE)
}
\arguments{
\item{update}{(logical) Whether to update the dataset list.}
}
\value{
A tibble.
}
\description{
List the dataset available at Open Data Jawa Barat.
}
